pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rates-provider"
version = "0.1.0"
description = "Exchange rates provider aggregator."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"aiogram>=3.4,<4.0",
	"python-dotenv>=1.0,<2.0",
]

[project.scripts]
rates-provider = "app.bootstrap.entrypoint:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]

[tool.mypy]
python_version = "3.13"
strict = true
pretty = true
show_error_codes = true
warn_unused_configs = true

[tool.pytest.ini_options]
pythonpath = ["src"]

[[tool.mypy.overrides]]
module = ["pytest"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["aiogram", "aiogram.*"]
ignore_missing_imports = true